Improve features & add labels
